update msvc build to sync with new sofia-sip.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@5844 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ libs/sofia-sip/libsofia-sip-ua/nta/nta_internal.h
@@ -146,7 +146,7 @@ struct nta_agent_s
   size_t                sa_maxsize;
   
   /** Maximum size of outgoing UDP requests */
-  size_t                sa_udp_mtu;
+  unsigned              sa_udp_mtu;
 
   /** SIP T1 - initial interval of retransmissions (500 ms) */
   unsigned              sa_t1;
